requirements.txt
Django==5.1.5
djangorestframework==3.15.2
PyPDF2==3.0.1
gunicorn==23.0.0
python-magic==0.4.27
django-cors-headers==4.6.0
packaging==24.2
asgiref==3.8.1
sqlparse==0.5.3
whitenoise==6.6.0
img2pdf==0.4.4
pikepdf==8.13.0